pkg: add cashaddr to index move cashaddr to cashaddr-browser for future C implementation
