+++ lib/bcstring.js
@@ -8,3 +8,4 @@
 
 exports.base58 = require('./base58');
 exports.bech32 = require('./bech32');
+exports.cashaddr = require('./cashaddr');
