+++ package.json
@@ -47,7 +47,8 @@
   "gypfile": true,
   "browser": {
     "./lib/base58": "./lib/base58-browser.js",
-    "./lib/bech32": "./lib/bech32-browser.js"
+    "./lib/bech32": "./lib/bech32-browser.js",
+    "./lib/cashaddr": "./lib/cashaddr-browser.js"
   },
   "browserify": {
     "transform": [
